- Bugfix: sticky windows forgot to stop their fade-in timers - We should have a base WindowController class which handles fade-in/out

+++ Display Plugins/Bubbles/GrowlBubblesWindowController.h
@@ -19,7 +19,7 @@
 	SEL				_action;
 	id				_target;
 	id				_representedObject;
-	short			_displayTime;
+	double			_displayTime;
 }
 
 + (GrowlBubblesWindowController *) bubble;
